Menu pembayaran

+++ app/Http/Controllers/SeragamDetailController.php
@@ -5,7 +5,9 @@ namespace App\Http\Controllers;
 use App\Models\Seragam;
 use App\Models\SeragamDetail;
 use App\Models\Supplier;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Inertia\Inertia;
 
 class SeragamDetailController extends Controller
@@ -15,13 +17,18 @@ class SeragamDetailController extends Controller
      */
     public function index()
     {
-        $supplier = Supplier::all();
-        $seragam = SeragamDetail::with('seragam')->paginate(100);
-        return Inertia::render('SeragamDetail/Index', [
-            'title' => "Daftar Ukuran Seragam",
-            'seragam' => $seragam,
-            'supplier' => $supplier,
-        ]);
+        $user = User::find(Auth::user()->id);
+        if ($user->hasRole('admin')) {
+            $supplier = Supplier::all();
+            $seragam = SeragamDetail::with('seragam')->paginate(100);
+            return Inertia::render('SeragamDetail/Index', [
+                'title' => "Daftar Ukuran Seragam",
+                'seragam' => $seragam,
+                'supplier' => $supplier,
+            ]);
+        } else {
+            return Inertia::render('Error/404');
+        }
     }
 
     /**
@@ -29,11 +36,16 @@ class SeragamDetailController extends Controller
      */
     public function create()
     {
-        $seragam = Seragam::All();
-        return Inertia::render('SeragamDetail/Create', [
-            'title' => "Tambah Ukuran Seragam",
-            'seragam' => $seragam,
-        ]);
+        $user = User::find(Auth::user()->id);
+        if ($user->hasRole('admin')) {
+            $seragam = Seragam::All();
+            return Inertia::render('SeragamDetail/Create', [
+                'title' => "Tambah Ukuran Seragam",
+                'seragam' => $seragam,
+            ]);
+        } else {
+            return Inertia::render('Error/404');
+        }
     }
 
     /**
@@ -41,19 +53,24 @@ class SeragamDetailController extends Controller
      */
     public function store(Request $request)
     {
-        $validator = $request->validate([
-            'ukuran' => 'required',
-            'seragam_id' => 'required'
-        ], [
-            'ukuran.required' => "Ukuran harus diisi",
-            'seragam_id.required' => "Data Seragam harus diisi",
-        ]);
-        $seragam = new SeragamDetail();
-        $seragam->ukuran = $request->ukuran;
-        $seragam->seragam_id = $request->seragam_id;
-        $seragam->stok = 0;
-        $seragam->save();
-        return to_route('seragam-detail.index');
+        $user = User::find(Auth::user()->id);
+        if ($user->hasRole('admin')) {
+            $validator = $request->validate([
+                'ukuran' => 'required',
+                'seragam_id' => 'required'
+            ], [
+                'ukuran.required' => "Ukuran harus diisi",
+                'seragam_id.required' => "Data Seragam harus diisi",
+            ]);
+            $seragam = new SeragamDetail();
+            $seragam->ukuran = $request->ukuran;
+            $seragam->seragam_id = $request->seragam_id;
+            $seragam->stok = 0;
+            $seragam->save();
+            return to_route('seragam-detail.index');
+        } else {
+            return Inertia::render('Error/404');
+        }
     }
 
     /**
@@ -61,18 +78,23 @@ class SeragamDetailController extends Controller
      */
     public function show($id)
     {
-        $seragam = SeragamDetail::with('seragam')->join('seragam', 'seragam.id', '=', 'seragam_detail.seragam_id')->where('kategori', $id)->paginate(100);
-        if ($id == 1) {
-            $unit = " PG";
-        } elseif ($id == 2) {
-            $unit = " TK";
-        } elseif ($id == 3) {
-            $unit = " SD";
+        $user = User::find(Auth::user()->id);
+        if ($user->hasRole('admin')) {
+            $seragam = SeragamDetail::with('seragam')->join('seragam', 'seragam.id', '=', 'seragam_detail.seragam_id')->where('kategori', $id)->paginate(100);
+            if ($id == 1) {
+                $unit = " PG";
+            } elseif ($id == 2) {
+                $unit = " TK";
+            } elseif ($id == 3) {
+                $unit = " SD";
+            }
+            return Inertia::render('SeragamDetail/Index', [
+                'title' => "Daftar Ukuran Seragam" . $unit,
+                'seragam' => $seragam,
+            ]);
+        } else {
+            return Inertia::render('Error/404');
         }
-        return Inertia::render('SeragamDetail/Index', [
-            'title' => "Daftar Ukuran Seragam" . $unit,
-            'seragam' => $seragam,
-        ]);
     }
 
     /**
@@ -80,13 +102,18 @@ class SeragamDetailController extends Controller
      */
     public function edit($id)
     {
-        $seragam = Seragam::All();
-        $seragamDetail = SeragamDetail::findOrFail($id);
-        return Inertia::render('SeragamDetail/Edit', [
-            'title' => "Update Seragam " . $seragamDetail->seragam->nama_seragam,
-            'seragam' => $seragam,
-            'seragamDetail' => $seragamDetail,
-        ]);
+        $user = User::find(Auth::user()->id);
+        if ($user->hasRole('admin')) {
+            $seragam = Seragam::All();
+            $seragamDetail = SeragamDetail::findOrFail($id);
+            return Inertia::render('SeragamDetail/Edit', [
+                'title' => "Update Seragam " . $seragamDetail->seragam->nama_seragam,
+                'seragam' => $seragam,
+                'seragamDetail' => $seragamDetail,
+            ]);
+        } else {
+            return Inertia::render('Error/404');
+        }
     }
 
     /**
@@ -94,18 +121,23 @@ class SeragamDetailController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $validator = $request->validate([
-            'ukuran' => 'required',
-            'seragam_id' => 'required'
-        ], [
-            'ukuran.required' => "Ukuran harus diisi",
-            'seragam_id.required' => "Data Seragam harus diisi",
-        ]);
-        $seragam = SeragamDetail::findOrFail($id);
-        $seragam->ukuran = $request->ukuran;
-        $seragam->seragam_id = $request->seragam_id;
-        $seragam->update();
-        return to_route('seragam-detail.index');
+        $user = User::find(Auth::user()->id);
+        if ($user->hasRole('admin')) {
+            $validator = $request->validate([
+                'ukuran' => 'required',
+                'seragam_id' => 'required'
+            ], [
+                'ukuran.required' => "Ukuran harus diisi",
+                'seragam_id.required' => "Data Seragam harus diisi",
+            ]);
+            $seragam = SeragamDetail::findOrFail($id);
+            $seragam->ukuran = $request->ukuran;
+            $seragam->seragam_id = $request->seragam_id;
+            $seragam->update();
+            return to_route('seragam-detail.index');
+        } else {
+            return Inertia::render('Error/404');
+        }
     }
 
     /**
@@ -113,8 +145,13 @@ class SeragamDetailController extends Controller
      */
     public function destroy($id)
     {
-        $seragam = SeragamDetail::findOrFail($id);
-        $seragam->delete();
-        return redirect()->back()->with('message', 'Data berhasil dihapus');
+        $user = User::find(Auth::user()->id);
+        if ($user->hasRole('admin')) {
+            $seragam = SeragamDetail::findOrFail($id);
+            $seragam->delete();
+            return redirect()->back()->with('message', 'Data berhasil dihapus');
+        } else {
+            return Inertia::render('Error/404');
+        }
     }
 }
