+++ app/Http/Controllers/SupplierController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Supplier;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Inertia\Inertia;
 
 class SupplierController extends Controller
@@ -13,11 +15,16 @@ class SupplierController extends Controller
      */
     public function index()
     {
-        $supplier = Supplier::paginate(100);
-        return Inertia::render('Supplier/Index', [
-            'title' => "Daftar Supplier",
-            'supplier' => $supplier,
-        ]);
+        $user = User::find(Auth::user()->id);
+        if ($user->hasRole('admin')) {
+            $supplier = Supplier::paginate(100);
+            return Inertia::render('Supplier/Index', [
+                'title' => "Daftar Supplier",
+                'supplier' => $supplier,
+            ]);
+        } else {
+            return Inertia::render('Error/404');
+        }
     }
 
     /**
@@ -25,9 +32,14 @@ class SupplierController extends Controller
      */
     public function create()
     {
-        return Inertia::render('Supplier/Create', [
-            'title' => "Tambah Supplier",
-        ]);
+        $user = User::find(Auth::user()->id);
+        if ($user->hasRole('admin')) {
+            return Inertia::render('Supplier/Create', [
+                'title' => "Tambah Supplier",
+            ]);
+        } else {
+            return Inertia::render('Error/404');
+        }
     }
 
     /**
@@ -35,24 +47,29 @@ class SupplierController extends Controller
      */
     public function store(Request $request)
     {
-        $validator = $request->validate([
-            'nama_supplier' => 'required',
-            'penanggung_jawab' => 'required',
-            'alamat' => 'required',
-            'telepon' => 'required'
-        ], [
-            'nama_supplier.required' => "Nama supplier harus diisi",
-            'penanggung_jawab.required' => "Kategori harus diisi",
-            'alamat.required' => "Nama supplier harus diisi",
-            'telepon.required' => "Kategori harus diisi",
-        ]);
-        $supplier = new Supplier();
-        $supplier->nama_supplier = $request->nama_supplier;
-        $supplier->penanggung_jawab = $request->penanggung_jawab;
-        $supplier->alamat = $request->alamat;
-        $supplier->telepon = $request->telepon;
-        $supplier->save();
-        return to_route('supplier.index');
+        $user = User::find(Auth::user()->id);
+        if ($user->hasRole('admin')) {
+            $validator = $request->validate([
+                'nama_supplier' => 'required',
+                'penanggung_jawab' => 'required',
+                'alamat' => 'required',
+                'telepon' => 'required'
+            ], [
+                'nama_supplier.required' => "Nama supplier harus diisi",
+                'penanggung_jawab.required' => "Kategori harus diisi",
+                'alamat.required' => "Nama supplier harus diisi",
+                'telepon.required' => "Kategori harus diisi",
+            ]);
+            $supplier = new Supplier();
+            $supplier->nama_supplier = $request->nama_supplier;
+            $supplier->penanggung_jawab = $request->penanggung_jawab;
+            $supplier->alamat = $request->alamat;
+            $supplier->telepon = $request->telepon;
+            $supplier->save();
+            return to_route('supplier.index');
+        } else {
+            return Inertia::render('Error/404');
+        }
     }
 
     /**
@@ -68,11 +85,16 @@ class SupplierController extends Controller
      */
     public function edit($id)
     {
-        $supplier = Supplier::findOrFail($id);
-        return Inertia::render('Supplier/Edit', [
-            'title' => "Edit Data Supplier",
-            'supplier' => $supplier,
-        ]);
+        $user = User::find(Auth::user()->id);
+        if ($user->hasRole('admin')) {
+            $supplier = Supplier::findOrFail($id);
+            return Inertia::render('Supplier/Edit', [
+                'title' => "Edit Data Supplier",
+                'supplier' => $supplier,
+            ]);
+        } else {
+            return Inertia::render('Error/404');
+        }
     }
 
     /**
@@ -80,24 +102,29 @@ class SupplierController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $validator = $request->validate([
-            'nama_supplier' => 'required',
-            'penanggung_jawab' => 'required',
-            'alamat' => 'required',
-            'telepon' => 'required'
-        ], [
-            'nama_supplier.required' => "Nama supplier harus diisi",
-            'penanggung_jawab.required' => "Kategori harus diisi",
-            'alamat.required' => "Nama supplier harus diisi",
-            'telepon.required' => "Kategori harus diisi",
-        ]);
-        $supplier = Supplier::findOrFail($id);
-        $supplier->nama_supplier = $request->nama_supplier;
-        $supplier->penanggung_jawab = $request->penanggung_jawab;
-        $supplier->alamat = $request->alamat;
-        $supplier->telepon = $request->telepon;
-        $supplier->update();
-        return to_route('supplier.index');
+        $user = User::find(Auth::user()->id);
+        if ($user->hasRole('admin')) {
+            $validator = $request->validate([
+                'nama_supplier' => 'required',
+                'penanggung_jawab' => 'required',
+                'alamat' => 'required',
+                'telepon' => 'required'
+            ], [
+                'nama_supplier.required' => "Nama supplier harus diisi",
+                'penanggung_jawab.required' => "Kategori harus diisi",
+                'alamat.required' => "Nama supplier harus diisi",
+                'telepon.required' => "Kategori harus diisi",
+            ]);
+            $supplier = Supplier::findOrFail($id);
+            $supplier->nama_supplier = $request->nama_supplier;
+            $supplier->penanggung_jawab = $request->penanggung_jawab;
+            $supplier->alamat = $request->alamat;
+            $supplier->telepon = $request->telepon;
+            $supplier->update();
+            return to_route('supplier.index');
+        } else {
+            return Inertia::render('Error/404');
+        }
     }
 
     /**
@@ -105,8 +132,13 @@ class SupplierController extends Controller
      */
     public function destroy($id)
     {
-        $supplier = Supplier::findOrFail($id);
-        $supplier->delete();
-        return redirect()->back()->with('message', 'Data berhasil dihapus');
+        $user = User::find(Auth::user()->id);
+        if ($user->hasRole('admin')) {
+            $supplier = Supplier::findOrFail($id);
+            $supplier->delete();
+            return redirect()->back()->with('message', 'Data berhasil dihapus');
+        } else {
+            return Inertia::render('Error/404');
+        }
     }
 }
